% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abpm_count.R
\name{abpm_count}
\alias{abpm_count}
\title{Count ABPM readings}
\usage{
abpm_count(data)
}
\arguments{
\item{data}{a \code{data.frame} produced by \link{read_bpm}.}
}
\value{
an integer vector (number of readings while awake and asleep)
}
\description{
Count ABPM readings
}
