% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_interrogate.R
\name{bp_interrogate}
\alias{bp_interrogate}
\title{Check for unexpected BP measurements}
\usage{
bp_interrogate(data)
}
\arguments{
\item{data}{a data.frame object from \code{read_bpm()}}
}
\description{
Check for unexpected BP measurements
}
