% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbpm_count.R
\name{hbpm_count}
\alias{hbpm_count}
\title{Count HBPM readings}
\usage{
hbpm_count(data, max_per_day = 4)
}
\arguments{
\item{data}{a \code{data.frame} produced by \link{read_bpm}.}

\item{max_per_day}{maximum number of readings per day. For example,
if \code{max_per_day} is 4 and there are 5 readings that day, then only
4 of those 5 count for that day.}
}
\value{
an integer (number of HBPM readings)
}
\description{
Count HBPM readings
}
\details{
\itemize{
\item so for example, 5 readings on one day counts as 4 readings)
}
}
