% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressureCounter-package.R
\docType{package}
\name{pressureCounter-package}
\alias{pressureCounter}
\alias{pressureCounter-package}
\title{pressureCounter: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcjaeger/pressureCounter}
  \item Report bugs at \url{https://github.com/bcjaeger/pressureCounter/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
