% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abpm.R
\name{read_bpm}
\alias{read_bpm}
\title{Read and clean BP monitoring data}
\usage{
read_bpm(fpath)
}
\arguments{
\item{fpath}{path to the abpm file}
}
\value{
a data.frame with bp values
}
\description{
Read and clean BP monitoring data
}
